+++ PerfectXL.EPPlus.UnitTests/FormulaParsing/Excel/Functions/TimeStringParserTests.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Text;
-using System.Collections.Generic;
-using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OfficeOpenXml.FormulaParsing.Excel.Functions.DateTime;
 
@@ -13,7 +10,7 @@ namespace EPPlusTest.Excel.Functions
         private double GetSerialNumber(int hour, int minute, int second)
         {
             var secondsInADay = 24d * 60d * 60d;
-            return ((double)hour * 60 * 60 + (double)minute * 60 + (double)second) / secondsInADay;
+            return ((double)hour * 60 * 60 + (double)minute * 60 + second) / secondsInADay;
         }
 
         [TestMethod]
