+++ PerfectXL.EPPlus.UnitTests/FormulaParsing/IntegrationTests/BuiltInFunctions/DateAndTimeFunctionsTests.cs
@@ -1,15 +1,8 @@
 ﻿using System;
-using System.Text;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OfficeOpenXml;
 using OfficeOpenXml.FormulaParsing;
-using OfficeOpenXml.FormulaParsing.ExpressionGraph;
 using FakeItEasy;
-using System.IO;
-using System.Threading;
 
 namespace EPPlusTest.FormulaParsing.IntegrationTests.BuiltInFunctions
 {
@@ -83,7 +76,7 @@ namespace EPPlusTest.FormulaParsing.IntegrationTests.BuiltInFunctions
         [TestMethod]
         public void TimeShouldReturnCorrectResult()
         {
-            var expectedResult = ((double)(12 * 60 * 60 + 13 * 60 + 14))/((double)(24 * 60 * 60));
+            var expectedResult = (12 * 60 * 60 + 13 * 60 + 14) / ((double)(24 * 60 * 60));
             var result = _parser.Parse("Time(12, 13, 14)");
             Assert.AreEqual(expectedResult, result);
         }
